Refactor query-based cumulative calculations
