+++ backend/server/graphql/schema.py
@@ -5,29 +5,14 @@ from functools import partial
 
 import graphene
 from django.utils import timezone
-from django.db.models import (
-    Count,
-    Case,
-    When,
-    Value,
-    IntegerField,
-    Subquery,
-    OuterRef,
-    Max,
-    Min,
-    Func,
-    F,
-    Sum,
-    Window,
-    Avg,
-    QuerySet,
-)
+from django.db.models import QuerySet, Count
 from mypy_extensions import TypedDict
 import pandas as pd
 import numpy as np
 
-from server.models import Prediction, Match, MLModel, TeamMatch
+from server.models import Prediction, Match, MLModel
 from server.types import RoundMetrics
+from server.graphql.calculations import cumulative_metrics_query
 from .types import (
     SeasonType,
     PredictionType,
@@ -173,49 +158,6 @@ def _calculate_cumulative_margin_difference(data_frame: pd.DataFrame):
     )
 
 
-def _calculate_cumulative_accuracy():
-    return Window(
-        expression=Avg("tip_point"),
-        partition_by=F("ml_model_id"),
-        order_by=F("match__start_date_time").asc(),
-    )
-
-
-def _calculate_cumulative_correct():
-    return Window(
-        expression=Sum("tip_point"),
-        partition_by=F("ml_model_id"),
-        order_by=F("match__start_date_time").asc(),
-    )
-
-
-def _calculate_absolute_margin_difference():
-    return Case(
-        When(
-            is_correct=True,
-            then=Func(F("predicted_margin") - F("match__margin"), function="ABS"),
-        ),
-        default=(F("predicted_margin") + F("match__margin")),
-        output_field=IntegerField(),
-    )
-
-
-def _get_match_winner_name():
-    return Subquery(
-        TeamMatch.objects.filter(match_id=OuterRef("match_id"))
-        .order_by("-score")
-        .values_list("team__name")[:1]
-    )
-
-
-def _calculate_tip_points():
-    return Case(
-        When(is_correct=True, then=Value(1)),
-        default=Value(0),
-        output_field=IntegerField(),
-    )
-
-
 class Query(graphene.ObjectType):
     """Base GraphQL Query type that contains all queries and their resolvers."""
 
@@ -342,43 +284,17 @@ class Query(graphene.ObjectType):
             .first()
         )
 
-        metric_values = (
-            Prediction.objects.filter(
-                match__start_date_time__year=max_match.year,
-                ml_model__used_in_competitions=True,
-                # We don't want to include unplayed matches, which would impact
-                # mean-based metrics like accuracy and MAE
-                match__start_date_time__lt=timezone.localtime(),
-            )
-            .select_related("ml_model", "match")
-            .order_by("match__start_date_time")
-            .annotate(
-                tip_point=_calculate_tip_points(),
-                match__winner__name=_get_match_winner_name(),
-                match__margin=(
-                    Max("match__teammatch__score") - Min("match__teammatch__score")
-                ),
-                absolute_margin_diff=_calculate_absolute_margin_difference(),
-                cumulative_correct_count=_calculate_cumulative_correct(),
-                cumulative_accuracy=_calculate_cumulative_accuracy(),
-            )
-            .values(
-                "match__start_date_time",
-                "match__start_date_time__year",
-                "match__round_number",
-                "ml_model__name",
-                "ml_model__used_in_competitions",
-                "predicted_margin",
-                "predicted_win_probability",
-                "predicted_winner__name",
-                "match__winner__name",
-                "match__margin",
-                "absolute_margin_diff",
-                "cumulative_correct_count",
-                "cumulative_accuracy",
-                "match__id",
-                "ml_model__is_principle",
-            )
+        prediction_query_set = Prediction.objects.filter(
+            match__start_date_time__year=max_match.year,
+            ml_model__used_in_competitions=True,
+        )
+        additional_metric_values = [
+            "match__start_date_time__year",
+            "match__id",
+            "ml_model__is_principle",
+        ]
+        metric_values = cumulative_metrics_query(
+            prediction_query_set, additional_metric_values,
         )
 
         metrics_df = (
